Move selected example handling to header component
